\name{Country}
\docType{data}
\alias{Country}
\title{Countries in the Penn World Table}
\description{
  The mapping between countries' ID and countries' name
}
\source{
    Summers, R.  and  A.  Heston (1991) \dQuote{The Penn world table (mark 5): an exanded set of international comparaisons, 1950-1988}, \emph{Quarterly Journal of Economics }, \bold{29}, 229-256.
}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, chapter 5, 358-363.
}
\keyword{datasets}