\name{Greene}
\docType{data}
\alias{Greene}
\title{Cost and factor shares for  U.S. Electricity Producers}
\description{
  a cross-section from 1955

 \emph{number of observations} :  99

 \emph{observation} :  firms

   \emph{country} :  United States

}
\usage{data(Greene)}
\format{A dataframe containing :
  \describe{
    \item{ID}{observation}
    \item{TC}{total costs in millions of dollars}
    \item{Q}{output in billions of kilowatt hours}
    \item{PL}{price of labor}
    \item{PK}{price of capital}
    \item{PF}{price of fuels}
    \item{SL}{price of labor}
    \item{SK}{price of capital}
  }
}
\source{
    Christensen, L.  and  W. H.  Greene (1976) \dQuote{Economies of scale in U.S. electric power generation}, \emph{Journal of Political Economy }, \bold{84}, 655-676.
}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, chapter 1, 76-84.
}
\keyword{datasets}