\name{Grilic}
\docType{data}
\alias{Grilic}
\title{Wage Datas}
\description{
  a cross-section from 1980

 \emph{number of observations} :  758

 \emph{observation} :  individual

   \emph{country} :  United States

}
\usage{data(Grilic)}
\format{A dataframe containing :
  \describe{
    \item{RNS}{dummy for residency in the southern states (first observation)}
    \item{RNS80}{dummy for residency in the southern states (1980)}
    \item{MRT}{dummy for marital status (first observation)}
    \item{MRT80}{dummy for marital status (1980)}
    \item{SMSA}{dummy for residency in metropolan areas (first observation)}
    \item{SMSA80}{dummy for residency in metropolan areas (1980)}
    \item{MED}{mother's education in years}
    \item{IQ}{iQ score}
    \item{KWW}{score on the ``knowledge of the world of work'' test}
    \item{YEAR}{year of the observation}
    \item{AGE}{age (first observation)}
    \item{AGE80}{age (1980)}
    \item{S}{completed years of schooling (first observation)}
    \item{S80}{completed years of schooling (1980)}
    \item{EXPR}{experience in years (first observation)}
    \item{EXPR80}{experience in years (1980)}
    \item{TENURE}{tenure in years (first observation)}
    \item{TENURE80}{tenure in years (1980)}
    \item{LW}{log wage (first observation)}
    \item{LW80}{log wage (1980)}
  }
}
\source{
    Blackburn, M.  and   Neumark D. (1992) \dQuote{Unobserved ability, efficiency wages, and interindustry wage differentials}, \emph{Quarterly Journal of Economics }, \bold{107}, 1421-1436.
}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, chapter 3, 250-256.
}
\section{Original description}{\dQuote{758 observations at two points in time.}}
\keyword{datasets}