\name{Nerlove}
\docType{data}
\alias{Nerlove}
\title{Cost of 145 U.S. Electricity Producers}
\description{
  a cross-section from 1955

 \emph{number of observations} :  145

 \emph{observation} :  firms

   \emph{country} :  United States

}
\usage{data(Nerlove)}
\format{A dataframe containing :
  \describe{
    \item{TC}{total costs in millions of dollars}
    \item{Q}{output in billions of kilowatt hours}
    \item{PL}{price of labor}
    \item{PF}{price of fuels}
    \item{PK}{price of capital}
  }
}
\source{
    Nerlove, M. (1963) \emph{Returns to scale in electricity industry} \emph{in}  Christ, C. ed. (1963) \emph{Measurment in economics: studies in mathematical economics and econometrics in memory of Yehuda Grunfeld }, Standford, California, Standford University Press .
}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, chapter 1, 76-84,.
}
\keyword{datasets}