\name{LT}
\docType{data}
\alias{LT}
\title{Dollar Sterling Exchange Rate}
\description{
  yearly  observations from 1791 to 1990\cr

 \emph{number of observations} :  200 \cr
   \emph{observation} :  country \cr
   \emph{country} :  United Kingdown \cr
    }
\usage{data(LT)}
\format{A dataframe containing :
  \describe{
    \item{Year}{the Year of observation}
    \item{S}{dollar / sterling exchange rate}
    \item{USWPI}{uS wholesale price index, normalized to 100 for 1914}
    \item{UKWPI}{uS wholesale price index, normalized to 100 for 1914}
  }
}
\source{
    Lothian, J.  and  M.  Taylor (1996) ``Real exchange rate behavior: the recent float from the perspective of the past two centuries'', \emph{Journal of Political Economy }, \bold{104}, 488-509. \cr
}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, chapter 9, 613-621. \cr
}
\keyword{datasets}