\name{Mpyr}
\docType{data}
\alias{Mpyr}
\title{Money, National Product and Interest Rate}
\description{
  yearly  observations from 1900 to 1989

 \emph{number of observations} :  90

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Mpyr)}
\format{A dataframe containing :
  \describe{
    \item{m}{natural log of M1}
    \item{p}{natural log of the net national product price deflator}
    \item{y}{natural log of the net national product}
    \item{R}{the commercial paper rate in percent at an annual rate}
  }
}
\source{
    T Stock, J.  and   Watson M. (1999) \dQuote{Testing for common trends}, \emph{Journal of the American Statistical Association }, \bold{83}, 1097-1107.
}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, chapter 10, 665-667.
}
\keyword{datasets}